requirements.txt
google-cloud-aiplatform
google-generativeai
langgraph>=0.2.0
langchain>=0.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
colorama>=0.4.6
tqdm>=4.66.0
pillow>=10.0.0